remove a few more warnings

+++ libstdio/vfdprintf.c
@@ -10,6 +10,6 @@ static int __fwrite(void*ptr, size_t nmemb, int fd) {
 
 int vfdprintf(int fd, const char *format, va_list arg_ptr)
 {
-  struct arg_printf ap = { (void*)fd, (int(*)(void*,size_t,void*)) __fwrite };
+  struct arg_printf ap = { (void*)(long)fd, (int(*)(void*,size_t,void*)) __fwrite };
   return __v_printf(&ap,format,arg_ptr);
 }
